requirements.txt
Django>=5.1
djangorestframework>=3.15
psycopg2-binary>=2.9
requests>=2.31
WeasyPrint>=62
openpyxl>=3.1
matplotlib>=3.8
Pillow>=10.0
whitenoise>=6.5
gunicorn>=22.0
openai>=1.0
docxtpl>=0.20.0
python-docx>=1.1.0
